+++ src/main/java/de/gematik/demis/notification/builder/demis/fhir/notification/builder/infectious/laboratory/AnonymousCompositionBuilder.java
@@ -26,12 +26,15 @@ package de.gematik.demis.notification.builder.demis.fhir.notification.builder.in
  * #L%
  */
 
+import static de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants.RECEPTION_TIME_STAMP_TYPE;
+
 import de.gematik.demis.notification.builder.demis.fhir.notification.utils.DemisConstants;
 import java.util.List;
 import javax.annotation.Nonnull;
 import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.Composition;
 import org.hl7.fhir.r4.model.DiagnosticReport;
+import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.Identifier;
 import org.hl7.fhir.r4.model.Patient;
 import org.hl7.fhir.r4.model.PractitionerRole;
@@ -60,7 +63,7 @@ public class AnonymousCompositionBuilder extends NotificationLaboratoryDataBuild
     builder.setCompositionStatus(original.getStatus());
 
     builder.setTitle(original.getTitle());
-    builder.setDate(original.getDate());
+    builder.setDateTimeType(original.getDateElement());
 
     List<Composition.CompositionRelatesToComponent> relatesTo = original.getRelatesTo();
     if (!relatesTo.isEmpty()) {
@@ -79,6 +82,12 @@ public class AnonymousCompositionBuilder extends NotificationLaboratoryDataBuild
     builder.setLaboratoryReport(diagnosticReport);
     final Composition.SectionComponent section = original.getSectionFirstRep();
     builder.setSectionCode(section.getCode().getCodingFirstRep());
+
+    Extension extensionByUrl = original.getExtensionByUrl(RECEPTION_TIME_STAMP_TYPE);
+    if (extensionByUrl != null) {
+      builder.addExtension(
+          new Extension().setUrl(extensionByUrl.getUrl()).setValue(extensionByUrl.getValue()));
+    }
 
     return builder.build();
   }